flatten protomanager
