+++ packages/insomnia/src/network/grpc/__tests__/write-proto-file.test.ts
@@ -5,8 +5,8 @@ import mkdirp from 'mkdirp';
 import os from 'os';
 import path from 'path';
 
-import { globalBeforeEach } from '../../../../__jest__/before-each';
-import * as models from '../../../../models';
+import { globalBeforeEach } from '../../../__jest__/before-each';
+import * as models from '../../../models';
 import writeProtoFile from '../write-proto-file';
 
 describe('writeProtoFile', () => {
