+++ packages/insomnia/src/network/grpc/write-proto-file.ts
@@ -3,12 +3,12 @@ import mkdirp from 'mkdirp';
 import os from 'os';
 import path from 'path';
 
-import { database as db } from '../../../common/database';
-import type { BaseModel } from '../../../models';
-import * as models from '../../../models';
-import { isProtoDirectory, ProtoDirectory } from '../../../models/proto-directory';
-import { isProtoFile, ProtoFile } from '../../../models/proto-file';
-import { isWorkspace, Workspace } from '../../../models/workspace';
+import { database as db } from '../../common/database';
+import type { BaseModel } from '../../models';
+import * as models from '../../models';
+import { isProtoDirectory, ProtoDirectory } from '../../models/proto-directory';
+import { isProtoFile, ProtoFile } from '../../models/proto-file';
+import { isWorkspace, Workspace } from '../../models/workspace';
 
 const getProtoTempFileName = ({ _id, modified }: ProtoFile): string => `${_id}.${modified}.proto`;
